+++ app/app.py
@@ -1,4 +1,4 @@
 from src import app
 
 if __name__ == "__main__":
-    app.run(host="0.0.0.0", port=5000)
+  app.run(host="0.0.0.0", port=8080)
