+++ app/requirements.txt
@@ -1,0 +1,7 @@
+boto3==1.17.33
+Flask==1.1.2
+Flask-Cors==3.0.10
+opencv-contrib-python
+opencv-python
+Pillow==8.1.2
+uWSGI==2.0.19.1
