+++ app/uwsgi/uwsgi.ini
@@ -3,8 +3,7 @@ wsgi-file = /var/app/app.py
 callable = app
 master = true
 processes = 4
-socket = 127.0.0.1:8080
-chmod-socket = 666
+http=0.0.0.0:8080
 vacuum = true
 die-on-term = true
 py-autoreload = 1
