+++ docker-compose.yml
@@ -1,17 +1,23 @@
 version: '3'
 
 services:
-
-  flask:
-    build: .
+  app:
+    build: app
     restart: always
     ports:
-      - 80:80
+      - "8080:8080"
     environment:
       APP_ENV: development
       S3_BUCKET: image-processing-181562662531
     volumes:
-        - ./src:/var/app/src
+        - src:/var/app/src
+
+  nginx:
+    build: nginx
+    ports:
+      - "80:80"
+    depends_on:
+      - app
 
 volumes:
   src: {}
