+++ nginx/Dockerfile
@@ -1,0 +1,7 @@
+FROM nginx:latest
+
+COPY default.conf /etc/nginx/conf.d/
+
+EXPOSE 80
+
+CMD ["nginx", "-g daemon off;"]
