+++ nginx/default.conf
@@ -1,6 +1,6 @@
 server {
-    listen 80 default_server;
-    root /var/app;
+    listen 80;
+    server_name localhost;
 
     keepalive_timeout 600s;
     send_timeout 600s;
@@ -9,9 +9,9 @@ server {
     proxy_send_timeout 600s;
     client_body_timeout 600s;
 
+    client_max_body_size 10m;
+
     location / {
-        include uwsgi_params;
-        uwsgi_pass 127.0.0.1:8080;
-        uwsgi_read_timeout 600s;
+        proxy_pass http://app:8080;
     }
 }
