--- startup.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-cd /var/app || exit
-
-sleep 3
-
-uwsgi --ini uwsgi/uwsgi.ini &
-
-nginx -g "daemon off;"
